requirements.txt
Django
Pillow
phonenumbers
python-dotenv
django-cors-headers
djangorestframework
django-phonenumber-field
djangorestframework-simplejwt
